+++ packages/featureserver/src/index.js
@@ -3,7 +3,7 @@ const info = require('./info.js')
 module.exports = {
   route: require('./route.js'),
   serverInfo: info.serverInfo,
-  serviceInfo: info.serviceInfo,
   layerInfo: info.layerInfo,
+  layersInfo: info.layersInfo,
   query: require('./query.js')
 }