+++ packages/featureserver/test/route.js
@@ -8,9 +8,15 @@ const app = express()
 
 let data
 
+const serverHander = (req, res) => {
+  FeatureServer.route(req, res, {
+    description: 'test',
+    layers: [data, data]
+  })
+}
 const handler = (req, res) => FeatureServer.route(req, res, data)
 
-app.get('/FeatureServer', handler)
+app.get('/FeatureServer', serverHander)
 app.get('/FeatureServer/layers', handler)
 app.get('/FeatureServer/:layer', handler)
 app.get('/FeatureServer/:layer/:method', handler)
@@ -21,57 +27,63 @@ describe('Routing feature server requests', () => {
     data.name = 'Snow'
   })
 
+  describe('Server Info', () => {
+    it('should properly route and handle a server info request`', done => {
+      request(app)
+        .get('/FeatureServer?f=json')
+        .expect(res => {
+          res.body.serviceDescription.should.equal('test')
+          res.body.layers.length.should.equal(2)
+          Array.isArray(res.body.tables).should.equal(true)
+        })
+        .expect('Content-Type', /json/)
+        .expect(200, done)
+    })
+  })
+
   describe('Query', () => {
     it('should properly route and handle a query', done => {
       request(app)
-      .get('/FeatureServer/0/query?f=json&where=1%3D1')
-      .expect(res => {
-        res.body.features[1].attributes.OBJECTID.should.equal(1)
-        res.body.features.length.should.equal(417)
-      })
-      .expect('Content-Type', /json/)
-      .expect(200, done)
+        .get('/FeatureServer/0/query?f=json&where=1%3D1')
+        .expect(res => {
+          res.body.features[1].attributes.OBJECTID.should.equal(1)
+          res.body.features.length.should.equal(417)
+        })
+        .expect('Content-Type', /json/)
+        .expect(200, done)
     })
   })
 
   describe('Layer Info', () => {
     it('should properly route and handle a layer info request`', done => {
       request(app)
-      .get('/FeatureServer/0?f=json')
-      .expect(res => {
-        res.body.type.should.equal('Feature Layer')
-        res.body.name.should.equal('Snow')
-        res.body.fields.filter(f => { return f.name === 'OBJECTID' }).length.should.equal(1)
-        should.exist(res.body.extent)
-      })
-      .expect('Content-Type', /json/)
-      .expect(200, done)
-    })
-  })
-
-  describe('Server Info', () => {
-    it('should properly route and handle a server info`', done => {
-      request(app)
-      .get('/FeatureServer?f=json')
-      .expect(res => {
-        res.body.layers.length.should.equal(1)
-      })
-      .expect('Content-Type', /json/)
-      .expect(200, done)
+        .get('/FeatureServer/0?f=json')
+        .expect(res => {
+          res.body.type.should.equal('Feature Layer')
+          res.body.name.should.equal('Snow')
+          res.body.fields
+            .filter(f => {
+              return f.name === 'OBJECTID'
+            })
+            .length.should.equal(1)
+          should.exist(res.body.extent)
+        })
+        .expect('Content-Type', /json/)
+        .expect(200, done)
     })
   })
 
   describe('Layers', () => {
     it('should properly route and handle a layers request`', done => {
       request(app)
-      .get('/FeatureServer/layers?f=json')
-      .expect(res => {
-        res.body.layers.length.should.equal(1)
-        res.body.tables.length.should.equal(0)
-        res.body.layers[0].name.should.equal('Snow')
-      })
-      .expect('Content-Type', /json/)
-      .expect(200, done)
+        .get('/FeatureServer/layers?f=json')
+        .expect(res => {
+          res.body.layers.length.should.equal(1)
+          res.body.tables.length.should.equal(0)
+          res.body.layers[0].name.should.equal('Snow')
+        })
+        .expect('Content-Type', /json/)
+        .expect(200, done)
     })
   })
 })