+++ packages/featureserver/CHANGELOG.md
@@ -2,7 +2,7 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
-## Unreleased
+## [2.1.0] - 2017-06-02
 ### Added
 * Support for passing a server extent e.g. [[-180, -90], [180, 90]]
 * Support passed geometryType passed in layer metadata
@@ -112,6 +112,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ## [1.0.0] - 2017-01-25
 * Initial Release
 
+[2.1.0]: https://github.com/featureserver/featureserver/compare/v2.0.0...v2.1.0
 [2.0.0]: https://github.com/featureserver/featureserver/compare/v2.0.0...v1.3.4
 [1.3.4]: https://github.com/featureserver/featureserver/compare/v1.3.3...v1.3.4
 [1.3.3]: https://github.com/featureserver/featureserver/compare/v1.3.2...v1.3.3